+++ packages/bundlekit-bundler-rollup/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bundlekit/bundler-rollup
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [a19cb50]
+  - @bundlekit/shared-utils@0.0.6
+
 ## 0.0.7
 
 ### Patch Changes